+++ code-pushup.config.ts
@@ -11,6 +11,7 @@ import {
   packageJsonPerformanceGroupRef,
   packageJsonPlugin,
 } from './dist/examples/plugins';
+import coveragePlugin from './dist/packages/plugin-coverage';
 import eslintPlugin, {
   eslintConfigFromNxProjects,
 } from './dist/packages/plugin-eslint';
@@ -48,7 +49,34 @@ const config: CoreConfig = {
 
   plugins: [
     await eslintPlugin(await eslintConfigFromNxProjects()),
-
+    coveragePlugin({
+      reports: [
+        {
+          resultsPath: 'coverage/cli/unit-tests/lcov.info',
+          pathToProject: 'packages/cli',
+        },
+        {
+          resultsPath: 'coverage/core/unit-tests/lcov.info',
+          pathToProject: 'packages/core',
+        },
+        {
+          resultsPath: 'coverage/models/unit-tests/lcov.info',
+          pathToProject: 'packages/models',
+        },
+        {
+          resultsPath: 'coverage/utils/unit-tests/lcov.info',
+          pathToProject: 'packages/utils',
+        },
+        {
+          resultsPath: 'coverage/plugin-eslint/unit-tests/lcov.info',
+          pathToProject: 'packages/plugin-eslint',
+        },
+        {
+          resultsPath: 'coverage/plugin-coverage/unit-tests/lcov.info',
+          pathToProject: 'packages/plugin-coverage',
+        },
+      ],
+    }),
     fileSizePlugin({
       directory: './dist/examples/react-todos-app',
       pattern: /\.js$/,
@@ -79,6 +107,30 @@ const config: CoreConfig = {
       title: 'Code style',
       refs: [
         { type: 'group', plugin: 'eslint', slug: 'suggestions', weight: 1 },
+      ],
+    },
+    {
+      slug: 'code-coverage',
+      title: 'Code coverage',
+      refs: [
+        {
+          type: 'audit',
+          plugin: 'coverage',
+          slug: 'function-coverage',
+          weight: 1,
+        },
+        {
+          type: 'audit',
+          plugin: 'coverage',
+          slug: 'branch-coverage',
+          weight: 1,
+        },
+        {
+          type: 'audit',
+          plugin: 'coverage',
+          slug: 'line-coverage',
+          weight: 1,
+        },
       ],
     },
     {